Added stub SvnHandler using a proxy.
This is based on Rack::StreamingProxy.  It creates a proxy on initialization
that rewrites URLs and then delegates to it on each request.  This now closely
matches the GitHandler code.  It still needs the configuration to be handled
and still needs authz code in the SvnHandler#call method.  This commit just
sets up the proxy structure, which will eventually be configurable to point to
a mod_dav_svn server.