Added notification area.
The new notification area is a place to display messages and popup forms.  It
displays system messages (renders shared/_system_messages), the flash
notifications (renders shared/_flash) and the :forms content.  The idea is to
have a fixed-position popup area that is not displayed until there is relevant
content.

So far, system messages aren't defined and there is just a stub partial in
place.  Flash messages are working but ugly; JS is in place to display the
messages and age-off notices and successes after 10 seconds.  Using a partial
to display flash makes it easy to render the flash for async JS calls so it is
appropriate to use for async form submission results.

Forms can be added to the area by using content_for :forms.  Forms should start
off with CSS class "hide".