Adding render_marker.
render_marker is a new ApplicationController method that sets up a marker page
and renders it.  This is supposed to allow us to write static documentation and
information pages (e.g., faq and about) in marker format, stored in git.  The
view it uses is shared/marker and marker docs should be stored in public/docs.

About and faq pages now use render_marker and empty files are in docs.